% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{buildPackage}
\alias{buildPackage}
\title{\t buildPackage}
\usage{
buildPackage(package, pull = build, build = check, check = cran,
  cran = FALSE, add = build, push = cran, install = build)
}
\arguments{
\item{package}{'character'}

\item{pull}{'logical'}

\item{build}{'logical'}

\item{check}{'logical'}

\item{cran}{'logical'}

\item{add}{'logical'}

\item{push}{'logical'}

\item{install}{'logical'
\t@description Build a package.
\t@import devtools
\t@import Rcpp}
}
\description{
\t buildPackage
}
