% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{detachPackage}
\alias{detachPackage}
\title{detachPackage}
\usage{
detachPackage(package, character.only = FALSE)
}
\arguments{
\item{package}{a 'character' describing a package to unload.}

\item{character.only}{a 'logical' flag, whether to skip deparsing/}
}
\description{
detachPackage
}
