% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{findPackages}
\alias{findPackages}
\title{\t findPackages
\t@description Provide package directories}
\usage{
findPackages(path = ".")
}
\arguments{
\item{path}{'character'
\t@export}
}
\description{
\t findPackages
\t@description Provide package directories
}
