% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitScripts.R, R/gitScripts_REMOTE_12702.R
\name{gitAdd}
\alias{gitAdd}
\alias{gitAdd}
\title{gitAdd}
\usage{
gitAdd(filelist)

gitAdd(filelist)
}
\arguments{
\item{filelist}{'character'}

\item{filelist}{'character'}
}
\description{
Git add, commit and push

Git add, commit and push
}
