% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitScripts.R, R/gitScripts_REMOTE_12702.R
\name{gitClone}
\alias{gitClone}
\alias{gitClone}
\title{gitClone}
\usage{
gitClone(url)

gitClone(url)
}
\arguments{
\item{url}{the url to clone}

\item{url}{the url to clone}
}
\description{
gitClone

gitClone
}
