% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitScripts.R, R/gitScripts_REMOTE_12702.R
\name{gitCommit}
\alias{gitCommit}
\alias{gitCommit}
\title{gitCommit}
\usage{
gitCommit(comment)

gitCommit(comment)
}
\arguments{
\item{comment}{'character'}

\item{comment}{'character'}
}
\description{
gitCommit

gitCommit
}
