% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitScripts.R, R/gitScripts_REMOTE_12702.R
\name{gitFetch}
\alias{gitFetch}
\alias{gitFetch}
\title{gitFetch}
\usage{
gitFetch(branch = NULL, checkout = TRUE, merge = checkout)

gitFetch(branch = NULL, checkout = TRUE, merge = checkout)
}
\arguments{
\item{branch}{remote branch to fetch}

\item{branch}{remote branch to fetch}
}
\description{
gitFetch

gitFetch
}
