% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitScripts.R, R/gitScripts_REMOTE_12702.R
\name{gitForcePush}
\alias{gitForcePush}
\alias{gitForcePush}
\title{gitForcePush}
\usage{
gitForcePush(branch = NULL)

gitForcePush(branch = NULL)
}
\arguments{
\item{branch}{a 'character' label of the branch to use}

\item{branch}{a 'character' label of the branch to use}
}
\description{
gitForcePush

gitForcePush
}
