% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitScripts.R, R/gitScripts_REMOTE_12702.R
\name{gitRebase}
\alias{gitRebase}
\alias{gitRebase}
\title{gitRebase}
\usage{
gitRebase(from)

gitRebase(from)
}
\arguments{
\item{from}{identifier to rebase upon.}

\item{tool}{mergetool to use}

\item{from}{identifier to rebase upon.}

\item{tool}{mergetool to use}
}
\description{
gitRebase

gitRebase
}
