% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitScripts.R, R/gitScripts_REMOTE_12702.R
\name{ls.git.files}
\alias{ls.git.files}
\alias{ls.git.files}
\title{ls.git.files}
\usage{
ls.git.files(pattern = ".*")

ls.git.files(pattern = ".*")
}
\arguments{
\item{pattern}{'character' a regexp to filter output}

\item{pattern}{'character' a regexp to filter output}
}
\description{
ls.git.files

ls.git.files
}
