% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\docType{package}
\name{buildPackage-package}
\title{build-package}
\description{
Routines for development, building and testing.
}
\author{
\strong{Maintainer}: Dale \email{dale@piratepress.org} [copyright holder]

}
