% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{buildPackage}
\alias{buildPackage}
\title{buildPackage}
\usage{
buildPackage(package, pull = build, build = check || push, check = cran,
  cran = FALSE, add = build, push = TRUE, install = build)
}
\arguments{
\item{package}{'character'}

\item{pull}{'logical'}

\item{build}{'logical'}

\item{check}{'logical'}

\item{cran}{'logical'}

\item{add}{'logical'}

\item{push}{'logical'}

\item{install}{'logical'}
}
\description{
Build a package.
}
