% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{buildPackage}
\alias{buildPackage}
\title{buildPackage}
\usage{
buildPackage(package, pull = build, build = check, check = as.cran,
  clean = push, as.cran = FALSE, push = check, install = build)
}
\description{
Build a package.
}
