% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\docType{package}
\name{buildPackage}
\alias{buildPackage}
\alias{buildPackage-package}
\title{What the package does (short line)}
\usage{
buildPackage(package)
}
\description{
More about what it does (maybe more than one line)

Build a package.
}

