% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{buildPackage}
\alias{buildPackage}
\title{buildPackage}
\usage{
buildPackage(package, pull = TRUE, check = TRUE, clean = TRUE,
  as.cran = FALSE, push = FALSE, install = TRUE)
}
\description{
Build a package.
}

