% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPackage.R
\name{buildPackage}
\alias{buildPackage}
\title{buildPackage}
\usage{
buildPackage(
  package,
  fetch = build,
  build = TRUE,
  check = cran,
  cran = FALSE,
  add = commit,
  commit = push,
  push = FALSE,
  install = build
)
}
\arguments{
\item{package}{'character'}

\item{fetch}{'logical'}

\item{build}{'logical'}

\item{check}{'logical'}

\item{cran}{'logical'}

\item{add}{'logical'}

\item{commit}{'logical'}

\item{push}{'logical'}

\item{install}{'logical'}
}
\description{
Build a package.
}
