% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitScripts.R
\name{gitExpunge}
\alias{gitExpunge}
\title{gitExpunge}
\usage{
gitExpunge(fileName, n = 200)
}
\arguments{
\item{fileName}{a 'character' vector of filenames to expunge from repo}

\item{n}{max number to handle}
}
\description{
gitExpunge
}
