% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitScripts.R
\name{gitFetch}
\alias{gitFetch}
\title{gitFetch}
\usage{
gitFetch(branch = NULL, merge = TRUE)
}
\arguments{
\item{branch}{'character' remote branch to fetch}

\item{merge}{'logical'}
}
\description{
gitFetch
}
